Simplify discrete random variable interface, add covariance support
